refactor: rename TextFieldHelpers files
